Make use of glib types in libtinypoker.
git-svn-id: http://svn.gna.org/svn/pokersource/trunk@5585 c4f34931-41c7-4605-b615-9bb9f479c57e

+++ tinypoker/libtinypoker/src/test/testRankPlayers.c
@@ -27,7 +27,7 @@ int main()
 	ipp_card *Ks, *Kd, *Kh, *Kc;
 	ipp_card *Qs, *Qh, *Qc;
 	ipp_card *Jh, *Jc, *Jd;
-	int *result;
+	guint8 *result;
 
 	ipp_player *player1, *player2, *player3, *player4;
 	ipp_table *table;
